requirements.txt
camel-oasis>=0.1.0
click>=8.1.0
pydantic>=2.6.0
rich>=13.7.0
python-dotenv>=1.0.0
openai>=1.30.0
pip-system-certs>=4.0; sys_platform == "win32"
